@@ -1,5 +1,5 @@
 buildscript {
-    ext.kotlin_version = '1.2.21'
+    ext.kotlin_version = '1.3.41'
 
     repositories {
         mavenCentral()
@@ -11,25 +11,40 @@ buildscript {
 }
 
 plugins {
-    id "org.jetbrains.kotlin.jvm" version "1.2.21"
+    id "org.jetbrains.kotlin.jvm" version "1.3.41"
 }
 
 apply plugin: 'kotlin'
 
 dependencies {
     compile "org.jetbrains.kotlin:kotlin-stdlib-jdk8:$kotlin_version"
-    testCompile 'junit:junit:4.12'
+
+    testImplementation('org.assertj:assertj-core:3.13.2')
+    testImplementation('org.junit.jupiter:junit-jupiter:5.5.1')
+    testImplementation('org.junit.jupiter:junit-jupiter-engine:5.5.1')
+    testImplementation('org.junit.jupiter:junit-jupiter-params:5.5.1')
+
 }
+
 repositories {
     mavenCentral()
 }
+
 compileKotlin {
     kotlinOptions {
         jvmTarget = "1.8"
     }
 }
+
 compileTestKotlin {
     kotlinOptions {
         jvmTarget = "1.8"
+    }
+}
+
+test {
+    useJUnitPlatform()
+    testLogging {
+        events "passed", "skipped", "failed"
     }
 }
